ocr/requirements_ocr.txt
pytesseract==0.3.10
easyocr==1.7.1
transformers==4.38.2
python-doctr==0.8.1
pdf2image==1.17.0
Pillow==10.2.0
torch==2.2.0
numpy==1.26.4
tabulate==0.9.0
tqdm==4.66.2
tf2onnx
tensorflow
